// seungjae/api-server/src/sample/types.ts
export type Timeframe =
	| '1m'
	| '3m'
	| '5m'
	| '15m'
	| '30m'
	| '60m'
	| '1H'
	| '4H'
	| '1D'
	| '1W'
	| '1M';